keeping all models in the models folder, integrating the get directions to the view bus times page
