+++ stopBApp/models/__init__.py
@@ -1,2 +1,6 @@
 from .guest import Guest
 from .authenticated import Authenticated
+from .location import Location
+from .busline import BusLine
+from .savedbusline import SavedBusLine
+from .detour import Detour
